BlockPopover: Remove __unstableCoverTarget in favour of BlockPopoverCover
